feat : add code to create simulation of Conway's game of life.
create Cell class with x and y attributes as cell's coordinates. Override equal and hashcode functions to use cell in hash set to know easy if a cell with x and y coordinates is already in a set. Both are need for all HashSet operation

Create ConwayGrid class as the grid of conway's game life. we only save the useful content - alive cells - to save memory and decrease time complexity. We use to hashSet to get state at time t to compute state at t+1.

Create ConwaySimulator as the simulator, implementing Simulable interface. Draw function reset and draw each current alive cell.

create a test file TestConwaySimulator to run a simulation with gliders to know if the sphere grid works.